Bug fix Added Optional to returned values in methods will work with DB

+++ src/main/java/net/juicy/api/utils/util/number/NumberUtil.java
@@ -15,7 +15,7 @@ public class NumberUtil {
         if (min >= max)
             return 0;
 
-        return ThreadLocalRandom.current().nextInt((int) (max - min)) + min;
+        return ThreadLocalRandom.current().nextInt(max - min) + min;
 
     }
